test/refactor: move node aggregations into own folders to make room for edge tests
