+++ packages/graphql/tests/integration/aggregations/where/node/datetime.int.test.ts
@@ -20,13 +20,11 @@
 import { Driver } from "neo4j-driver";
 import { graphql } from "graphql";
 import { generate } from "randomstring";
-import neo4j from "../../neo4j";
-import { Neo4jGraphQL } from "../../../../src/classes";
+import neo4j from "../../../neo4j";
+import { Neo4jGraphQL } from "../../../../../src/classes";
 
-describe("aggregations-where-bigint", () => {
+describe("aggregations-where-node-datetime", () => {
     let driver: Driver;
-
-    const bigInt = "2147483647";
 
     beforeAll(async () => {
         driver = await neo4j();
@@ -36,13 +34,13 @@ describe("aggregations-where-bigint", () => {
         await driver.close();
     });
 
-    test("should return posts where a like BigInt is EQUAL to", async () => {
+    test("should return posts where a like DateTime is EQUAL to", async () => {
         const session = driver.session();
 
         const typeDefs = `
             type User {
                 testString: String!
-                someBigInt: BigInt
+                someDateTime: DateTime!
             }
           
             type Post {
@@ -56,23 +54,25 @@ describe("aggregations-where-bigint", () => {
             readable: true,
         });
 
+        const someDateTime = new Date();
+
         const neoSchema = new Neo4jGraphQL({ typeDefs });
 
         try {
             await session.run(
                 `
-                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someBigInt: ${bigInt}})
+                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someDateTime: dateTime("${someDateTime.toISOString()}")})
                     CREATE (:Post {testString: "${testString}"})
                 `
             );
 
             const query = `
                 {
-                    posts(where: { testString: "${testString}", likesAggregate: { node: { someBigInt_EQUAL: ${bigInt} } } }) {
+                    posts(where: { testString: "${testString}", likesAggregate: { node: { someDateTime_EQUAL: "${someDateTime.toISOString()}" } } }) {
                         testString
                         likes {
                             testString
-                            someBigInt
+                            someDateTime
                         }
                     }
                 }
@@ -93,7 +93,7 @@ describe("aggregations-where-bigint", () => {
             expect((gqlResult.data as any).posts).toEqual([
                 {
                     testString,
-                    likes: [{ testString, someBigInt: bigInt }],
+                    likes: [{ testString, someDateTime: someDateTime.toISOString() }],
                 },
             ]);
         } finally {
@@ -101,13 +101,13 @@ describe("aggregations-where-bigint", () => {
         }
     });
 
-    test("should return posts where a like BigInt is GT than", async () => {
+    test("should return posts where a like DateTime is GT than", async () => {
         const session = driver.session();
 
         const typeDefs = `
             type User {
                 testString: String!
-                someBigInt: BigInt
+                someDateTime: DateTime!
             }
           
             type Post {
@@ -121,26 +121,27 @@ describe("aggregations-where-bigint", () => {
             readable: true,
         });
 
-        const someBigInt = `${bigInt}1`;
-        const someBigIntGt = bigInt.substring(0, bigInt.length - 1);
+        const someDateTime = new Date();
+        const someDateTimeGT = new Date();
+        someDateTimeGT.setDate(someDateTimeGT.getDate() - 1);
 
         const neoSchema = new Neo4jGraphQL({ typeDefs });
 
         try {
             await session.run(
                 `
-                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someBigInt: ${someBigInt}})
+                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someDateTime: datetime("${someDateTime.toISOString()}")})
                     CREATE (:Post {testString: "${testString}"})
                 `
             );
 
             const query = `
                 {
-                    posts(where: { testString: "${testString}", likesAggregate: { node: { someBigInt_GT: ${someBigIntGt} } } }) {
+                    posts(where: { testString: "${testString}", likesAggregate: { node: { someDateTime_GT: "${someDateTimeGT.toISOString()}" } } }) {
                         testString
                         likes {
                             testString
-                            someBigInt
+                            someDateTime
                         }
                     }
                 }
@@ -161,7 +162,7 @@ describe("aggregations-where-bigint", () => {
             expect((gqlResult.data as any).posts).toEqual([
                 {
                     testString,
-                    likes: [{ testString, someBigInt }],
+                    likes: [{ testString, someDateTime: someDateTime.toISOString() }],
                 },
             ]);
         } finally {
@@ -169,13 +170,13 @@ describe("aggregations-where-bigint", () => {
         }
     });
 
-    test("should return posts where a like BigInt is GTE than", async () => {
+    test("should return posts where a like DateTime is GTE than", async () => {
         const session = driver.session();
 
         const typeDefs = `
             type User {
                 testString: String!
-                someBigInt: BigInt
+                someDateTime: DateTime!
             }
           
             type Post {
@@ -189,23 +190,25 @@ describe("aggregations-where-bigint", () => {
             readable: true,
         });
 
+        const someDateTime = new Date();
+
         const neoSchema = new Neo4jGraphQL({ typeDefs });
 
         try {
             await session.run(
                 `
-                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someBigInt: ${bigInt}})
+                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someDateTime: datetime("${someDateTime.toISOString()}")})
                     CREATE (:Post {testString: "${testString}"})
                 `
             );
 
             const query = `
                 {
-                    posts(where: { testString: "${testString}", likesAggregate: { node: { someBigInt_GTE: ${bigInt} } } }) {
+                    posts(where: { testString: "${testString}", likesAggregate: { node: { someDateTime_GTE: "${someDateTime.toISOString()}" } } }) {
                         testString
                         likes {
                             testString
-                            someBigInt
+                            someDateTime
                         }
                     }
                 }
@@ -226,7 +229,7 @@ describe("aggregations-where-bigint", () => {
             expect((gqlResult.data as any).posts).toEqual([
                 {
                     testString,
-                    likes: [{ testString, someBigInt: bigInt }],
+                    likes: [{ testString, someDateTime: someDateTime.toISOString() }],
                 },
             ]);
         } finally {
@@ -234,13 +237,13 @@ describe("aggregations-where-bigint", () => {
         }
     });
 
-    test("should return posts where a like BigInt is LT than", async () => {
+    test("should return posts where a like DateTime is LT than", async () => {
         const session = driver.session();
 
         const typeDefs = `
             type User {
                 testString: String!
-                someBigInt: BigInt
+                someDateTime: DateTime!
             }
           
             type Post {
@@ -254,25 +257,27 @@ describe("aggregations-where-bigint", () => {
             readable: true,
         });
 
-        const someBigIntLT = `${bigInt}1`;
+        const someDateTime = new Date();
+        const someDateTimeLT = new Date();
+        someDateTimeLT.setDate(someDateTimeLT.getDate() + 1);
 
         const neoSchema = new Neo4jGraphQL({ typeDefs });
 
         try {
             await session.run(
                 `
-                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someBigInt: ${bigInt}})
+                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someDateTime: datetime("${someDateTime.toISOString()}")})
                     CREATE (:Post {testString: "${testString}"})
                 `
             );
 
             const query = `
                 {
-                    posts(where: { testString: "${testString}", likesAggregate: { node: { someBigInt_LT: ${someBigIntLT} } } }) {
+                    posts(where: { testString: "${testString}", likesAggregate: { node: { someDateTime_LT: "${someDateTimeLT.toISOString()}" } } }) {
                         testString
                         likes {
                             testString
-                            someBigInt
+                            someDateTime
                         }
                     }
                 }
@@ -293,7 +298,7 @@ describe("aggregations-where-bigint", () => {
             expect((gqlResult.data as any).posts).toEqual([
                 {
                     testString,
-                    likes: [{ testString, someBigInt: bigInt }],
+                    likes: [{ testString, someDateTime: someDateTime.toISOString() }],
                 },
             ]);
         } finally {
@@ -301,13 +306,13 @@ describe("aggregations-where-bigint", () => {
         }
     });
 
-    test("should return posts where a like BigInt is LTE than", async () => {
+    test("should return posts where a like DateTime is LTE than", async () => {
         const session = driver.session();
 
         const typeDefs = `
             type User {
                 testString: String!
-                someBigInt: BigInt
+                someDateTime: DateTime!
             }
           
             type Post {
@@ -321,23 +326,25 @@ describe("aggregations-where-bigint", () => {
             readable: true,
         });
 
+        const someDateTime = new Date();
+
         const neoSchema = new Neo4jGraphQL({ typeDefs });
 
         try {
             await session.run(
                 `
-                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someBigInt: ${bigInt}})
+                    CREATE (:Post {testString: "${testString}"})<-[:LIKES]-(:User {testString: "${testString}", someDateTime: datetime("${someDateTime.toISOString()}")})
                     CREATE (:Post {testString: "${testString}"})
                 `
             );
 
             const query = `
                 {
-                    posts(where: { testString: "${testString}", likesAggregate: { node: { someBigInt_LTE: ${bigInt} } } }) {
+                    posts(where: { testString: "${testString}", likesAggregate: { node: { someDateTime_LTE: "${someDateTime.toISOString()}" } } }) {
                         testString
                         likes {
                             testString
-                            someBigInt
+                            someDateTime
                         }
                     }
                 }
@@ -358,7 +365,7 @@ describe("aggregations-where-bigint", () => {
             expect((gqlResult.data as any).posts).toEqual([
                 {
                     testString,
-                    likes: [{ testString, someBigInt: bigInt }],
+                    likes: [{ testString, someDateTime: someDateTime.toISOString() }],
                 },
             ]);
         } finally {
