+++ packages/graphql/tests/integration/aggregations/where/node/float.int.test.ts
@@ -20,10 +20,10 @@
 import { Driver } from "neo4j-driver";
 import { graphql } from "graphql";
 import { generate } from "randomstring";
-import neo4j from "../../neo4j";
-import { Neo4jGraphQL } from "../../../../src/classes";
+import neo4j from "../../../neo4j";
+import { Neo4jGraphQL } from "../../../../../src/classes";
 
-describe("aggregations-where-float", () => {
+describe("aggregations-where-node-float", () => {
     let driver: Driver;
 
     beforeAll(async () => {
