+++ packages/graphql/tests/tck/tck-test-files/cypher/aggregations/where/node/int.md
@@ -1,4 +1,4 @@
-# Cypher Aggregations where with Int
+# Cypher Aggregations where node with Int
 
 Tests for queries inside the relationship where aggregation arg using an Int type.
